Remove: Follows type 분리
